refactor : postImageConverter 삭제 후 postConverter로 이동 및 리팩토링 (#207)
