+++ src/main/java/umc/duckmelang/domain/member/converter/MemberProfileImageConverter.java
@@ -2,7 +2,6 @@ package umc.duckmelang.domain.member.converter;
 
 import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Component;
-import umc.duckmelang.domain.member.domain.Member;
 import umc.duckmelang.domain.member.domain.MemberProfileImage;
 import umc.duckmelang.domain.member.dto.profileImage.MemberProfileImageResponseDto;
 
@@ -50,5 +49,4 @@ public class MemberProfileImageConverter {
                 .changedStatus(changedStatus)
                 .build();
     }
-
 }
